💩Fixing relink issues. ✨Adding a debug rule. 🐛Fixing bugs related to linux os.

+++ srcs/ft_strtrim.c
@@ -3,14 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strtrim.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ldutriez <ldutriez@student.42.fr>          +#+  +:+       +#+        */
+/*   By: notcampeur <notcampeur@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/10 11:00:41 by ldutriez          #+#    #+#             */
-/*   Updated: 2020/02/24 14:25:49 by ldutriez         ###   ########.fr       */
+/*   Updated: 2021/10/01 12:28:00 by notcampeur       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/libft.h"
+#include "libft.h"
 
 static int		is_unwanted(char c, char const *set)
 {
@@ -54,7 +54,7 @@ static int		locate_the_end(char const *s1, char const *set)
 		end_s1--;
 		end = end_s1;
 	}
-	return (end == 0 ? ft_strlen(s1) : end);
+	return (end == 0 ? (int)ft_strlen(s1) : end);
 }
 
 char			*ft_strtrim(char const *s1, char const *set)
